Job is deleted from the UI
